Limit information character count to 180 characters
