+++ www/front_src/src/Resources/Details/tabs/Timeline/Event.tsx
@@ -25,6 +25,7 @@ import {
 import { getFormattedTime, getFormattedDateTime } from '../../../dateTime';
 import DowntimeChip from '../../../Chip/Downtime';
 import AcknowledgeChip from '../../../Chip/Acknowledge';
+import truncate from '../../../truncate';
 
 const types: Array<Type> = [
   {
@@ -80,16 +81,6 @@ interface Props {
 const Date = ({ event }: Props): JSX.Element => (
   <Typography variant="caption">{getFormattedTime(event.date)}</Typography>
 );
-
-const truncate = (content: string): string => {
-  const maxLength = 180;
-
-  if (gt(content.length, maxLength)) {
-    return `${content.substring(0, maxLength)}...`;
-  }
-
-  return content;
-};
 
 const Content = ({ event }: Props): JSX.Element => {
   const { content } = event;
